dev: add path option to places-and-screens script
This adds a `--path` option to the `places_and_screens.json` builder
script, allowing it to be pointed at any local copy of the screen
config rather than pulling it from one of the deployed environments.
This also enables the script to be run without AWS credentials.

If run without options, the script now prints usage information.

The script is also renamed for clarity, given it does not "fetch" the
places-and-screens config file, it "builds" it (though possibly using
data "fetched" from the Screens config).