fix: ignore lab sign in places script
Similar to how other screens can be marked `hidden_from_screenplay`,
this enables flagging specific PA/ESS signs that we don't want shown in
Screenplay. This fixes the issue that the output of `build_places.exs`
was no longer directly uploadable since the addition of the lab sign,
and had to be manually tweaked to remove it.