feat: support filtering PaMessages by route
Derives a mapping of route id to pa/ess sign ids from the "places and
screens" data.

Uses the mapping of routes to PA/ESS sign ids to expand the the sign
filtering to include all signs for all of the routes included in the
options passed.

Required adding a `route_ids` field to the `PaEssScreen` struct.