fix: re-enable review audio
The `useEffect` based updating of `audioState` was resetting the
`audioState` to "outdated" if the message text or phonetic audio were
different from their starting values.

This manually sets `audioState` in each place that should cause
`audioState` to become "outdated" instead. This is how this state was
updated prior to the `useEffect` based refactor.