chore: handle unauthorized state on new PA message
Normally a user without the PA message admin role would not have any UI
to navigate to this page anyway, but on the off-chance they get there
somehow, we can make it clear they're not authorized to use it. The API
endpoint to actually create a new PA message is secured by a server-side
role check.